datastores: refactor ProcessInfo to align with protobuf API
Refactor ProcessInfo struct to match protobuf naming and properly represent
containerized processes with separate namespace PIDs.

Changes:
- Rename EntityID (uint64) -> UniqueId (uint32) - matches hash output
- Rename PID/PPID -> HostPid/HostPpid - clarifies host namespace
- Add Pid/Ppid - namespace PIDs for containerized processes
- Add ParentUniqueId - efficient parent lookups
- Add ExitTime - process lifecycle tracking
- Update ProcessStore interface to use uint32 entityId consistently

Fixes dispatcher bug where both HostPid and Pid were set to ancestor.PID,
incorrectly representing containerized process ancestry.

+++ pkg/datastores/registry_test.go
@@ -23,15 +23,15 @@ type mockProcessStore struct {
 	mockDataStore
 }
 
-func (m *mockProcessStore) GetProcess(entityId uint64) (*datastores.ProcessInfo, bool) {
+func (m *mockProcessStore) GetProcess(entityId uint32) (*datastores.ProcessInfo, bool) {
 	return nil, false
 }
 
-func (m *mockProcessStore) GetChildProcesses(entityId uint64) ([]*datastores.ProcessInfo, error) {
+func (m *mockProcessStore) GetChildProcesses(entityId uint32) ([]*datastores.ProcessInfo, error) {
 	return nil, nil
 }
 
-func (m *mockProcessStore) GetAncestry(entityId uint64, maxDepth int) ([]*datastores.ProcessInfo, error) {
+func (m *mockProcessStore) GetAncestry(entityId uint32, maxDepth int) ([]*datastores.ProcessInfo, error) {
 	return nil, nil
 }
